# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NRCC — Node-RED Control Center API
  version: 1.0.0
  description: |
    REST API for the Node-RED Control Center (NRCC).

    ## Authentication

    Most endpoints require a short-lived Bearer JWT (≈15 min) obtained from
    `/api/auth/login` or `/api/auth/setup`. Refresh tokens are managed via an
    httpOnly cookie (`nrcc_refresh`) and rotated automatically on each call to
    `/api/auth/refresh`.

    ## Response Envelope

    Every JSON endpoint wraps its payload in a standard envelope:

    **Success:**
    ```json
    { "success": true, "data": <T>, "timestamp": "<RFC3339>" }
    ```

    **Error:**
    ```json
    { "success": false, "error": { "code": "ERROR_CODE", "message": "..." }, "timestamp": "<RFC3339>" }
    ```
  license:
    name: MIT
  contact:
    name: NRCC Team
servers:
  - url: /
    description: Same-origin (default deployment)
...
security: []
paths:
...
  /api/config:
    get:
      tags: [Config]
      operationId: getConfig
      summary: Get Node-RED configuration
      description: Returns the current Node-RED `settings.js` configuration as a structured object.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope_NodeRedConfig"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [Config]
      operationId: postConfig
      summary: Save Node-RED configuration
      description: "**Admin only.** Overwrites the Node-RED configuration."
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeRedConfig"
      responses:
        "200":
          description: Saved configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope_NodeRedConfig"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/system/info:
    get:
      tags: [System]
      operationId: getSystemInfo
      summary: Get system resource information
      description: Returns real-time CPU, memory, and disk statistics for the host.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: System information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope_SystemInfo"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/system/security-posture:
    get:
      tags: [System]
      operationId: getSecurityPosture
      summary: Get security posture chips
      description: |
        Admin-only. Returns the four boolean/count chips that back the
        Dashboard's SecurityPostureCard. The encryptionKeyConfigured chip
        surfaces the silent-degradation failure mode in issue #04:
        encrypted env vars written in clear when NRCC_ENCRYPTION_KEY is
        missing. Closes #676 item 2.
      security:
        - bearerAuth: []
        - adminOnly: []
      responses:
        "200":
          description: Security posture summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope_SecurityPosture"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ─── Backups ───────────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Short-lived JWT obtained from `/api/auth/login` or `/api/auth/setup`.
        Lifetime: ~15 minutes. Refresh using `POST /api/auth/refresh` (httpOnly
        cookie `nrcc_refresh`).
    adminOnly:
      type: apiKey
      in: header
      name: X-Admin-Only
      description: |
        Marker security scheme used in operations that require the admin role
        in addition to the bearer JWT. The endpoint rejects non-admin JWTs at
        runtime (RequireAdmin middleware); the schema simply documents the
        gate so consumers see which operations need elevated privileges.

  # ── Parameters ──────────────────────────────────────────────────────────────────
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
          enum:
            - INVALID_REQUEST
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - WEAK_PASSWORD
            - USERNAME_EXISTS
            - UPDATE_IN_PROGRESS
            - BACKUP_ERROR
            - FLOW_ERROR
            - ENV_ERROR
            - AUTH_FAILED
            - RATE_LIMITED
            - SETUP_REQUIRED
            - ALREADY_INITIALIZED
          example: INVALID_REQUEST
        message:
          type: string
          description: Human-readable error description
          example: Username and password are required
    ErrorEnvelope:
      type: object
      required: [success, error, timestamp]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          $ref: "#/components/schemas/ApiError"
        timestamp:
          type: string
          format: date-time
          description: RFC3339 UTC timestamp

    # ── Domain Models ──────────────────────────────────────────────────────────────
...
    NodeRedConfig:
      type: object
      description: Node-RED settings.js configuration
      properties:
        port:
          type: integer
          description: HTTP listener port
          example: 1880
        uiPort:
          type: integer
          description: Alias for port (used by frontend)
          example: 1880
        uiHost:
          type: string
        httpAdminRoot:
          type: string
          example: /
        httpNodeRoot:
          type: string
          example: /
        flowFile:
          type: string
          example: flows.json
        userDir:
          type: string
          example: ./data
        nodesDir:
          type: string
        credentialSecret:
          type: string
          description: Encryption secret for Node-RED credentials
        disableEditor:
          type: boolean
        adminAuth:
          $ref: "#/components/schemas/AdminAuth"
        projectsEnabled:
          type: boolean
        lang:
          type: string
          example: en-US
        functionLibraries:
          type: array
          items:
            $ref: "#/components/schemas/FunctionLibrary"
        envVars:
          type: array
          items:
            $ref: "#/components/schemas/EnvVar"
    AdminAuth:
      type: object
      required: [type]
      properties:
        type:
          type: string
          example: credentials
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminAuthUser"
    AdminAuthUser:
      type: object
      required: [username, password, permissions]
      properties:
        username:
          type: string
        password:
          type: string
        permissions:
          type: string
    FunctionLibrary:
      type: object
      required: [name]
      properties:
        name:
          type: string
        alias:
          type: string
        version:
          type: string
    EnvVar:
      type: object
      required: [key, value, type]
      properties:
        key:
          type: string
        value:
          type: string
        type:
          type: string
          enum: [string, number, boolean, secret]
          description: Variable type; `secret` values are stored encrypted
        description:
          type: string
        encrypted:
          type: boolean
...
    CpuInfo:
      type: object
      required: [usage, cores]
      properties:
        usage:
          type: number
          format: double
          description: CPU usage percentage (0–100)
        cores:
          type: integer
    MemoryInfo:
      type: object
      required: [total, free, used, usagePercent]
      properties:
        total:
          type: integer
          format: int64
          description: Total memory in bytes
        free:
          type: integer
          format: int64
        used:
          type: integer
          format: int64
        usagePercent:
          type: number
          format: double
    DiskInfo:
      type: object
      required: [total, free, used, usagePercent]
      properties:
        total:
          type: integer
          format: int64
          description: Total disk space in bytes
        free:
          type: integer
          format: int64
        used:
          type: integer
          format: int64
        usagePercent:
          type: number
          format: double
    SystemInfo:
      type: object
      required: [platform, arch, nodeVersion, hostname, uptime, cpu, memory, disk, nodeRedVersion, edgeMode]
      properties:
        platform:
          type: string
          example: linux
        arch:
          type: string
          example: amd64
        nodeVersion:
          type: string
          example: v20.11.0
        hostname:
          type: string
        uptime:
          type: integer
          format: int64
          description: Host uptime in seconds
        cpu:
          $ref: "#/components/schemas/CpuInfo"
        memory:
          $ref: "#/components/schemas/MemoryInfo"
        disk:
          $ref: "#/components/schemas/DiskInfo"
        nodeRedVersion:
          type: string
          example: 4.0.2
        edgeMode:
          type: boolean
          description: Whether NRCC runs in edge mode (EDGE_MODE). Default false.
          example: false

    # ── Backups ────────────────────────────────────────────────────────────────────
...
    SuccessEnvelope_NodeRedConfig:
      allOf:
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              enum: [true]
            timestamp:
              type: string
              format: date-time
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/NodeRedConfig"
...
    SuccessEnvelope_SystemInfo:
      allOf:
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              enum: [true]
            timestamp:
              type: string
              format: date-time
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SystemInfo"
...
    SecurityPosture:
      type: object
      required: [encryptionKeyConfigured, backupAccessAdminGated, activeRefreshSessions, totalAdmins, mfaEnrolledAdmins]
      properties:
        encryptionKeyConfigured:
          type: boolean
          description: |
            Whether NRCC_ENCRYPTION_KEY is set to a non-empty value. False
            means env vars flagged Encrypted are persisted in clear,
            which is the silent-degradation failure mode the Dashboard
            security posture card surfaces. See issue #04.
        backupAccessAdminGated:
          type: boolean
          description: |
            Whether GET /api/backups/{id}/download requires the admin
            role. Reflected here so the dashboard flags it if the gate
            is ever loosened. See issue #01.
        activeRefreshSessions:
          type: integer
          description: Count of non-revoked refresh sessions whose expiry is in the future.
        totalAdmins:
          type: integer
          description: Number of users with the admin role.
        mfaEnrolledAdmins:
          type: integer
          description: Subset of totalAdmins with TOTP enrolled.
    SuccessEnvelope_SecurityPosture:
      allOf:
        - type: object
          required: [success, data, timestamp]
          properties:
            success:
              type: boolean
              enum: [true]
            timestamp:
              type: string
              format: date-time
        - type: object
          properties:
            data:
              $ref: "#/components/schemas/SecurityPosture"

  # ── Reusable Responses ──────────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Bad request — invalid input or missing required fields
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: INVALID_REQUEST
              message: Username and password are required
            timestamp: "2026-05-24T14:00:00Z"
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authentication required
            timestamp: "2026-05-24T14:00:00Z"
    Forbidden:
      description: Authenticated but insufficient permissions (admin required)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
          example:
            success: false
            error:
              code: FORBIDDEN
              message: Admin access required
            timestamp: "2026-05-24T14:00:00Z"
...